pyproject.toml
[project]
name = "bombtask"
version = "0.1.0"
description = ""
authors = [
    {name = "Until",email = "kostyuchenko.work@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "redis (>=6.2.0,<7.0.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "aiohttp (>=3.12.14,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "anyio (>=4.9.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
